Encapsulated DownArrow animations into its own component
